migration: rack_migrate now takes two bounds like rack_crawl
It uses these two bounds for filtering its list of known commits to the window
between these two timestamps.
