+++ migration/mypy.ini
@@ -6,6 +6,7 @@ disallow_untyped_defs = True
 exclude = build
 
 files =
+    git_helpers,
     migration_helpers,
     ontology_changes,
     rack,
