+++ migration/ontology_changes/ontology_change.py
@@ -13,7 +13,6 @@ from dataclasses import dataclass
 import logging
 from abc import ABC, abstractmethod
 from typing import List
-import sys
 
 from colorama import Fore, Style
 from semtk import SemTKJSON
