+++ migration/rack_migrate/rack_migrate
@@ -15,11 +15,13 @@ import argparse
 import logging
 import sys
 from pathlib import Path
+from typing import Callable, List, NoReturn, Tuple
 
 import semtk
 
 from custom_formatter import stream_handler
-from ontology_changes.ontology_change import LOGGER_ID, stylize_file_name
+import git_helpers as git
+from ontology_changes.ontology_change import Commit, LOGGER_ID, stylize_file_name
 from rack.commits import commits_in_chronological_order
 
 
@@ -28,6 +30,21 @@ logger.addHandler(stream_handler)
 logger.propagate = False
 
 parser = argparse.ArgumentParser()
+
+parser.add_argument(
+    "--old-ref",
+    help="Old reference to migrate from",
+    type=str,
+    required=True,
+)
+
+parser.add_argument(
+    "--new-ref",
+    help="More recent reference to migrate up to",
+    type=str,
+    required=True,
+)
+
 
 parser.add_argument(
     "--from-folder",
@@ -56,16 +73,90 @@ parser.add_argument(
 
 args = parser.parse_args()
 
+
 try:
     logging.basicConfig(level=args.log_level)
 except ValueError:
     logger.error("Bad log level specified")
     sys.exit(1)
 
+
+old_commit_id = git.get_commit_id(args.old_ref)
+new_commit_id = git.get_commit_id(args.new_ref)
+
+
+def get_commit_by_id(
+    commits: List[Commit],
+    commit_id: str,
+    abort: Callable[[], NoReturn],
+) -> Tuple[int, Commit]:
+    try:
+        return next(
+            commit for commit in enumerate(commits) if commit[1].number == commit_id
+        )
+    except StopIteration as e:
+        print(e)
+        abort()
+
+
+def make_abort(message: str) -> Callable[[], NoReturn]:
+    def abort() -> NoReturn:
+        print(message)
+        sys.exit(1)
+
+    return abort
+
+
+def find_recorded_commits_between(
+    old_commit_id: str,
+    new_commit_id: str,
+) -> List[Commit]:
+
+    commit_ids_in_chronological_order: List[str] = []
+
+    # This callback will be called for all commits in the git history, from
+    # newer ones to older ones.  Therefore, prepending will yield commits in
+    # chronological order.
+    def on_commit(commit_id: str) -> None:
+        commit_ids_in_chronological_order.insert(0, commit_id)
+
+    git.traverse_commits(
+        new_commit=new_commit_id,
+        old_commit=old_commit_id,
+        on_commit=on_commit,
+    )
+
+    # Now all we need to do is to filter out all our recorded commits that are
+    # within this window.
+    return [
+        commit
+        for commit in commits_in_chronological_order
+        if any(
+            [
+                commit_id == commit.number
+                for commit_id in commit_ids_in_chronological_order
+            ]
+        )
+    ]
+
+
+commits_to_consider_in_chronological_order = find_recorded_commits_between(
+    old_commit_id=old_commit_id,
+    new_commit_id=new_commit_id,
+)
+
+if len(commits_to_consider_in_chronological_order) == 0:
+    print("We did not find any commits in the range you specified, aborting.")
+    sys.exit(1)
+
+
 if args.list_changes_only:
-    for commit in commits_in_chronological_order:
+
+    for commit in commits_to_consider_in_chronological_order:
         commit_number = commit.number
-        print(f"Commit {commit_number} https://github.com/ge-high-assurance/RACK/commit/{commit_number}")
+        print(
+            f"Commit {commit_number} https://github.com/ge-high-assurance/RACK/commit/{commit_number}"
+        )
         for change in commit.changes:
             print(f"  {change.text_description()}")
     sys.exit(0)
@@ -93,7 +184,7 @@ for json_file_path in from_folder.glob("*.json"):
     )
     with open(json_file_path) as json_file:
         my_json = semtk.SemTKJSON.parse_raw(json_file.read())
-        for commit in commits_in_chronological_order:
+        for commit in commits_to_consider_in_chronological_order:
             commit_number = commit.number
             logger.debug(f"Processing commit {commit_number}")
             for change in commit.changes:
